Did some cleanup and worked on a workaround for the odometry coordinates using the lefthand rule for angles.
