+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/testing/teleop/DriveToPositionTest.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opmodes;
+package org.firstinspires.ftc.teamcode.opmodes.testing.teleop;
 
 import com.arcrobotics.ftclib.command.CommandOpMode;
 import com.arcrobotics.ftclib.command.InstantCommand;
